refactor(userspace-dp): split event_stream/codec.rs into submodules
Split the 1165-LOC event_stream/codec.rs wire-codec monolith into a
codec/ module directory with four wire-format-concern submodules behind
a re-export shell (#4651). The former single file fused HA session-sync
frames, RT_FLOW telemetry frames, wire constants, and the decode path in
one contract surface, so any wire change forced review of unrelated frame
families together.

This is PURE code motion — no behavior change. The Rust compiler and the
existing codec_tests are the gate.

New layout under userspace-dp/src/event_stream/codec/:
  - wire.rs         message-type/flag constants, RT_FLOW event/action/
                    disposition byte values, header + address write
                    helpers, and the header-only stream-control frame
                    encoders (drain_complete / full_resync).
  - session_sync.rs the correctness-critical HA session-sync delta frames
                    (encode_session_open / encode_session_close) and the
                    close_flags helper.
  - rt_flow.rs      the RT_FLOW telemetry frames (dataplane.Event wire
                    shape), DataplaneEventKind, and DataplaneEventPayload.
  - decode.rs       the Go-parity RT_FLOW payload decoder plus its read
                    helpers.
  - mod.rs          re-export shell. EventFrame's struct definition and
                    its frame-introspection impl (as_bytes,
                    dataplane_event_payload, dataplane_event_kind,
                    is_session_sync, decode method) stay in the module
                    root so its pub(super) fields/methods keep
                    pub(in event_stream) visibility unchanged.

Public paths are preserved: mod.rs glob-re-exports every submodule so
every pre-split codec::<item> path (external consumers and the relocated
test module via `use super::*`) resolves unchanged. No caller outside the
codec module changed.

Byte-identity: the moved fn/impl/const/struct/enum bodies are unchanged
character-for-character. A visibility-normalized decl inventory is
identical before and after (74 fn/struct/enum/const declarations, same
signatures); a glue-stripped, visibility-normalized content multiset diff
shows only added comment/banner lines and zero code differences. The only
signature changes are visibility qualifiers: formerly module-private
constants and helpers that are now read across the split submodules or by
the relocated test module gained pub(super) (== pub(in codec), the exact
crate-internal scope they had as codec.rs module-private items — no wider
than before). The single impl EventFrame block became four (one per file
carrying EventFrame methods); impl DataplaneEventKind stays one.

codec_tests.rs moved into codec/ so its `#[path]` load and `use super::*`
still resolve the codec module; the external types its builders used to
reach through codec.rs's own top-level `use` lines are now imported
directly in the test module.

Validation:
  - cargo build: clean.
  - cargo test --release -- --test-threads=1 (FULL suite, serial):
    3854 passed / 0 failed / 2 ignored across 5 test binaries.
  - cargo test --release event_stream: 85 passed / 0 failed.

Closes #4651
