+++ userspace-dp/src/event_stream/codec/rt_flow.rs
@@ -1,156 +1,14 @@
-//! Wire codec for the event stream binary protocol.
+//! RT_FLOW telemetry frames (the canonical `dataplane.Event` wire shape) plus
+//! the dataplane-event kind enum and payload struct.
 //!
-//! Pure encoding/decoding functions with zero I/O — all frame construction
-//! happens on a stack-allocated `[u8; 256]` buffer.
+//! Split out of the former single-file `codec.rs` (#4651). These frames carry
+//! deny/screen/filter security events and the RT_FLOW SESSION_CREATE/CLOSE
+//! records consumed by the Go NetFlow/IPFIX exporters.
 
-use crate::afxdp::ForwardingDisposition;
-use crate::session::{SessionDecision, SessionDelta, SessionKey, SessionMetadata};
-use rustc_hash::FxHashMap;
 use std::net::IpAddr;
 
-// ---------------------------------------------------------------------------
-// Wire format constants
-// ---------------------------------------------------------------------------
-
-pub(crate) const FRAME_HEADER_SIZE: usize = 16;
-
-pub(crate) const MSG_SESSION_OPEN: u8 = 1;
-pub(crate) const MSG_SESSION_CLOSE: u8 = 2;
-#[allow(dead_code)]
-pub(crate) const MSG_SESSION_UPDATE: u8 = 3;
-pub(crate) const MSG_ACK: u8 = 4;
-pub(crate) const MSG_PAUSE: u8 = 5;
-pub(crate) const MSG_RESUME: u8 = 6;
-pub(crate) const MSG_DRAIN_REQUEST: u8 = 7;
-pub(crate) const MSG_DRAIN_COMPLETE: u8 = 8;
-pub(crate) const MSG_FULL_RESYNC: u8 = 9;
-pub(crate) const MSG_KEEPALIVE: u8 = 10;
-
-// #1379 codec-only security event foundation. Producer wiring must add
-// fixed-size non-blocking emission, rate limiting, and loss accounting.
-#[allow(dead_code)]
-pub(crate) const MSG_POLICY_DENY: u8 = 11;
-#[allow(dead_code)]
-pub(crate) const MSG_SCREEN_DROP: u8 = 12;
-#[allow(dead_code)]
-pub(crate) const MSG_FILTER_LOG: u8 = 13;
-
-// #2460: RT_FLOW SESSION_CLOSE on the raw dataplane-event channel. This
-// frame carries the canonical 144-byte `dataplane.Event` payload (the same
-// shape the Go `logging.DecodeRawEventRecord` / `ProcessRawEvent` parser
-// consumes for the deny/screen/filter frames above) with the event-type
-// byte set to RT_FLOW SESSION_CLOSE (2). It is what drives the Go
-// NetFlow/IPFIX session-close exporters in userspace mode. It is ADDITIVE
-// to — and entirely separate from — the minimal `MSG_SESSION_CLOSE` (type
-// 2) HA session-sync delta below; the two are emitted as a pair from one
-// close, never substituting for each other.
-pub(crate) const MSG_SESSION_CLOSE_RT_FLOW: u8 = 14;
-
-// #2508: RT_FLOW SESSION_CREATE on the raw dataplane-event channel. Same
-// 144-byte payload shape as the close frame but with the event-type byte set
-// to RT_FLOW SESSION_CREATE (1). Emitted ONLY for sessions admitted by a
-// policy with `then log session-init`; there is no flowexport consumer of
-// session opens, so this frame is gated at the producer (the caller checks
-// metadata.log_session_init). The Go logEvent path formats it as an
-// RT_FLOW_SESSION_CREATE syslog record.
-pub(crate) const MSG_SESSION_CREATE_RT_FLOW: u8 = 15;
-
-// #2749: 152 bytes (was 144). The trailing [144:152] block carries the
-// class-of-service / interface-attribution fields the NetFlow v9 + IPFIX
-// SESSION_CLOSE exporters re-introduce: [144] src ToS byte (DSCP<<2),
-// [145] accumulated TCP control bits, [146] flow direction (reserved, 0 —
-// deferred follow-up), [147] reserved, [148:152] egress ifindex (LE u32).
-// Only the SESSION_CLOSE frame populates these; every other frame leaves
-// [144:152] zero. The growth is purely ADDITIVE: the Go side keeps its
-// minimum-frame acceptance at the legacy 144 bytes (`rawEventWireSize`) and
-// reads [144:152] only when the frame actually carries them
-// (`len(data) >= 152`), so a new daemon still accepts an old (144-byte)
-// helper's frames and an old daemon ignores the trailing 8 bytes — both
-// rolling-upgrade directions are safe (#1961 both-sides wire discipline).
-//
-// #3056: [136:140] carries the admitting policy ID on the SESSION_CLOSE
-// frame, whose [44:48] policy_id slot is occupied by the #2853
-// created-subsec-nanos remainder and so cannot hold the policy ID the way the
-// SESSION_CREATE / deny / screen / filter frames do. [140:144] stays reserved
-// padding. All frame encoders emit this length; non-close frames leave
-// [136:152] zero.
-#[allow(dead_code)]
-pub(crate) const SECURITY_EVENT_PAYLOAD_SIZE: usize = 152;
-
-const RT_FLOW_AF_INET: u8 = 2;
-const RT_FLOW_AF_INET6: u8 = 10;
-// #2460: SESSION_CLOSE event-type byte. Must equal the Go
-// `eventTypeSessionClose` (pkg/logging/ringbuf.go) and
-// `dataplane.EventTypeSessionClose` (pkg/dataplane/types.go) value of 2,
-// so `eventTypeName(data[52])` resolves to the literal "SESSION_CLOSE" the
-// flowexport callbacks gate on.
-const RT_FLOW_EVENT_SESSION_CLOSE: u8 = 2;
-// #2508: SESSION_CREATE event-type byte. Must equal the Go
-// `eventTypeSessionOpen` (pkg/logging/ringbuf.go) and
-// `dataplane.EventTypeSessionOpen` value of 1 so `eventTypeName(data[52])`
-// resolves to "SESSION_OPEN", which the Go formatter renders as the
-// RT_FLOW_SESSION_CREATE syslog tag.
-const RT_FLOW_EVENT_SESSION_OPEN: u8 = 1;
-// #2508: per-policy SYSLOG log gate, written to the final payload byte
-// (offset 135). 1 == emit the per-policy RT_FLOW SYSLOG record on the Go
-// side; 0 == suppress it (the global flowexport close exporter is
-// unaffected).
-const RT_FLOW_LOG_SYSLOG: u8 = 1;
-const RT_FLOW_EVENT_POLICY_DENY: u8 = 3;
-const RT_FLOW_EVENT_SCREEN_DROP: u8 = 4;
-const RT_FLOW_EVENT_FILTER_LOG: u8 = 6;
-// #2460: SESSION_CLOSE reason byte. The HA close delta carries no
-// idle/FIN/RST/age discriminator, so the RT_FLOW close reports "none" (0).
-// A real close-reason (and the byte/packet counters + session duration this
-// frame currently reports as 0) is the per-session accounting work tracked
-// in #2501.
-const RT_FLOW_CLOSE_REASON_NONE: u8 = 0;
-// DENY/PERMIT are wire-format documentation pinned by codec_tests;
-// like REJECT below they have no non-test reader (#1826: exposed when
-// the stale module-level allow was removed).
-#[allow(dead_code)]
-const RT_FLOW_ACTION_DENY: u8 = 0;
-#[allow(dead_code)]
-const RT_FLOW_ACTION_PERMIT: u8 = 1;
-#[allow(dead_code)]
-const RT_FLOW_ACTION_REJECT: u8 = 2;
-
-/// Disposition encoding for the wire format.
-const DISP_FORWARD_CANDIDATE: u8 = 0;
-const DISP_LOCAL_DELIVERY: u8 = 1;
-const DISP_FABRIC_REDIRECT: u8 = 2;
-const DISP_POLICY_DENIED: u8 = 3;
-const DISP_NO_ROUTE: u8 = 4;
-const DISP_MISSING_NEIGHBOR: u8 = 5;
-const DISP_HA_INACTIVE: u8 = 6;
-const DISP_DISCARD_ROUTE: u8 = 7;
-const DISP_NEXT_TABLE_UNSUPPORTED: u8 = 8;
-
-// Flag bits for SessionOpen/Close
-pub(crate) const FLAG_FABRIC_REDIRECT: u8 = 1 << 0;
-pub(crate) const FLAG_FABRIC_INGRESS: u8 = 1 << 1;
-pub(crate) const FLAG_IS_REVERSE: u8 = 1 << 2;
-// #2785: carry the admitting policy's per-policy `then log` selection on
-// the HA session-sync open frame so a session that fails over to the peer
-// keeps emitting the same RT_FLOW SESSION_CREATE/CLOSE syslog records after
-// takeover. Additive bits on the existing flags byte: an old peer leaves
-// them clear (0), which decodes to "no per-policy log" — bit-identical to
-// pre-#2785 behavior (rolling-upgrade safe).
-pub(crate) const FLAG_LOG_SESSION_INIT: u8 = 1 << 3;
-pub(crate) const FLAG_LOG_SESSION_CLOSE: u8 = 1 << 4;
-// #4565: mark a NAT64 cross-address-family session so a peer-PROMOTED session
-// can rebuild its RFC 6146 reverse BIB after failover. A NAT64 forward flow is
-// keyed on the ORIGINAL IPv6 5-tuple, but its NAT decision rewrites to an IPv4
-// pool source (`snat_v4`) + IPv4 destination, and the reverse (v4->v6) reply is
-// keyed on `(server_v4 -> snat_v4, translated port)`. The synced generic NAT
-// address fields cannot carry a v4 source in a v6 session's 16-byte slot
-// unambiguously, so this flag tells the standby: (a) set `nat.nat64`, (b) read
-// the trailing `snat_v4` (below), (c) reconstruct the original v6 src/dst from
-// the forward KEY (they ARE `key.src_ip`/`key.dst_ip` for a NAT64 forward
-// session) and `dst_v4` from the /96-embedded low 32 bits of `key.dst_ip`.
-// Additive bit: an old peer leaves it clear (0) -> decodes to "not NAT64",
-// bit-identical to pre-#4565 (rolling-upgrade safe).
-pub(crate) const FLAG_NAT64: u8 = 1 << 5;
+use super::EventFrame;
+use super::wire::*;
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 pub(crate) enum DataplaneEventKind {
@@ -180,7 +38,7 @@ impl DataplaneEventKind {
         }
     }
 
-    fn from_msg_type(msg_type: u8) -> Option<Self> {
+    pub(super) fn from_msg_type(msg_type: u8) -> Option<Self> {
         match msg_type {
             MSG_POLICY_DENY => Some(Self::PolicyDeny),
             MSG_SCREEN_DROP => Some(Self::ScreenDrop),
@@ -191,7 +49,7 @@ impl DataplaneEventKind {
         }
     }
 
-    fn rt_flow_event_type(self) -> u8 {
+    pub(super) fn rt_flow_event_type(self) -> u8 {
         match self {
             Self::PolicyDeny => RT_FLOW_EVENT_POLICY_DENY,
             Self::ScreenDrop => RT_FLOW_EVENT_SCREEN_DROP,
@@ -209,7 +67,7 @@ impl DataplaneEventKind {
     // decoder. The budget RELEASE path on the I/O thread keys frames by their
     // msg_type byte via `from_msg_type` above, so SessionClose/SessionCreate
     // round-trip the per-kind queue budget correctly without this mapping.
-    fn from_rt_flow_event_type(event_type: u8) -> Option<Self> {
+    pub(super) fn from_rt_flow_event_type(event_type: u8) -> Option<Self> {
         match event_type {
             RT_FLOW_EVENT_POLICY_DENY => Some(Self::PolicyDeny),
             RT_FLOW_EVENT_SCREEN_DROP => Some(Self::ScreenDrop),
@@ -248,261 +106,7 @@ pub(crate) struct DataplaneEventPayload {
     pub(crate) timestamp_ns: u64,
 }
 
-// ---------------------------------------------------------------------------
-// EventFrame -- zero-allocation stack-buffered wire frame
-// ---------------------------------------------------------------------------
-
-/// Pre-serialized event frame ready for socket write.
-#[derive(Clone)]
-pub(crate) struct EventFrame {
-    pub(super) data: [u8; 256],
-    pub(super) len: u16,
-    pub(crate) seq: u64,
-}
-
 impl EventFrame {
-    /// Encode a SessionOpen (type 1) or SessionUpdate (type 3) frame.
-    pub(crate) fn encode_session_open(
-        seq: u64,
-        key: &SessionKey,
-        decision: &SessionDecision,
-        metadata: &SessionMetadata,
-        _zone_name_to_id: &FxHashMap<String, u16>,
-        fabric_redirect_sync: bool,
-    ) -> Self {
-        let mut buf = [0u8; 256];
-        let mut pos = FRAME_HEADER_SIZE; // skip header, fill later
-
-        // [0] AddrFamily
-        let is_v6 = key.addr_family == libc::AF_INET6 as u8;
-        buf[pos] = if is_v6 { 6 } else { 4 };
-        pos += 1;
-
-        // [1] Protocol
-        buf[pos] = key.protocol;
-        pos += 1;
-
-        // [2:4] SrcPort LE
-        buf[pos..pos + 2].copy_from_slice(&key.src_port.to_le_bytes());
-        pos += 2;
-
-        // [4:6] DstPort LE
-        buf[pos..pos + 2].copy_from_slice(&key.dst_port.to_le_bytes());
-        pos += 2;
-
-        // [6:8] NATSrcPort LE
-        let nat = &decision.nat;
-        buf[pos..pos + 2].copy_from_slice(&nat.rewrite_src_port.unwrap_or(0).to_le_bytes());
-        pos += 2;
-
-        // [8:10] NATDstPort LE
-        buf[pos..pos + 2].copy_from_slice(&nat.rewrite_dst_port.unwrap_or(0).to_le_bytes());
-        pos += 2;
-
-        // [10:14] OwnerRGID i32 LE
-        // #2467: widened from i16 to i32. Linux ifindexes (the sibling fields
-        // below) are a full `int` and wrap negative past 32767 on long-running
-        // systems with interface churn; owner_rg_id is widened with them to
-        // keep the record's three identity ints a uniform 32-bit width.
-        buf[pos..pos + 4].copy_from_slice(&(metadata.owner_rg_id as i32).to_le_bytes());
-        pos += 4;
-
-        // [14:18] EgressIfindex i32 LE (#2467)
-        buf[pos..pos + 4]
-            .copy_from_slice(&(decision.resolution.egress_ifindex as i32).to_le_bytes());
-        pos += 4;
-
-        // [18:22] TXIfindex i32 LE (#2467)
-        buf[pos..pos + 4].copy_from_slice(&(decision.resolution.tx_ifindex as i32).to_le_bytes());
-        pos += 4;
-
-        // [22:24] TunnelEndpointID u16 LE
-        buf[pos..pos + 2].copy_from_slice(&decision.resolution.tunnel_endpoint_id.to_le_bytes());
-        pos += 2;
-
-        // [24:26] TXVLANID u16 LE
-        buf[pos..pos + 2].copy_from_slice(&decision.resolution.tx_vlan_id.to_le_bytes());
-        pos += 2;
-
-        // [26] Flags
-        let mut flags: u8 = 0;
-        if fabric_redirect_sync
-            || decision.resolution.disposition == ForwardingDisposition::FabricRedirect
-        {
-            flags |= FLAG_FABRIC_REDIRECT;
-        }
-        if metadata.fabric_ingress {
-            flags |= FLAG_FABRIC_INGRESS;
-        }
-        if metadata.is_reverse {
-            flags |= FLAG_IS_REVERSE;
-        }
-        // #2785: per-policy log selection rides the open frame so the
-        // synced session logs identically on the peer after failover.
-        if metadata.log_session_init {
-            flags |= FLAG_LOG_SESSION_INIT;
-        }
-        if metadata.log_session_close {
-            flags |= FLAG_LOG_SESSION_CLOSE;
-        }
-        // #4565: signal a NAT64 cross-family session so the peer rebuilds the
-        // reverse BIB after failover (see FLAG_NAT64 doc + the trailing snat_v4).
-        if decision.nat.nat64 {
-            flags |= FLAG_NAT64;
-        }
-        buf[pos] = flags;
-        pos += 1;
-
-        // [27:29] IngressZoneID u16 LE (#3075: widened from u8).
-        // SessionMetadata.ingress_zone is u16; the event-stream now carries the
-        // full u16 so a stable name-hash zone id > 255 (#3075) round-trips
-        // instead of truncating. This is same-host same-version IPC — xpfd
-        // spawns the helper as a child and the #1917 STOP->FLIP->START kills the
-        // helper before the new daemon starts (socket unlinked+recreated, a
-        // FullResync drains the stream), so no frame straddles the width change
-        // and no record-version negotiation is required.
-        buf[pos..pos + 2].copy_from_slice(&metadata.ingress_zone.to_le_bytes());
-        pos += 2;
-
-        // [29:31] EgressZoneID u16 LE (#3075).
-        buf[pos..pos + 2].copy_from_slice(&metadata.egress_zone.to_le_bytes());
-        pos += 2;
-
-        // [31] Disposition u8
-        buf[pos] = encode_disposition(decision.resolution.disposition);
-        pos += 1;
-
-        // Addresses: 4 bytes each for v4, 16 bytes each for v6
-        pos = write_ip(&mut buf, pos, key.src_ip, is_v6);
-        pos = write_ip(&mut buf, pos, key.dst_ip, is_v6);
-        pos = write_ip_opt(&mut buf, pos, nat.rewrite_src, is_v6);
-        pos = write_ip_opt(&mut buf, pos, nat.rewrite_dst, is_v6);
-
-        // NeighborMAC [6 bytes]
-        if let Some(mac) = decision.resolution.neighbor_mac {
-            buf[pos..pos + 6].copy_from_slice(&mac);
-        }
-        pos += 6;
-
-        // SrcMAC [6 bytes]
-        if let Some(mac) = decision.resolution.src_mac {
-            buf[pos..pos + 6].copy_from_slice(&mac);
-        }
-        pos += 6;
-
-        // NextHop (4 or 16 bytes)
-        pos = write_ip_opt(&mut buf, pos, decision.resolution.next_hop, is_v6);
-
-        // #3301: the admitting policy's firewall metadata rides the open frame
-        // (trailing fields, additive within MSG_SESSION_OPEN) so a peer-promoted
-        // session is correctly attributed, counted, and aged after failover.
-        // The Go decoder length-skips these; an old daemon ignores them.
-        //
-        // [+0:+4] policy_id u32 LE (#3056)
-        buf[pos..pos + 4].copy_from_slice(&metadata.policy_id.to_le_bytes());
-        pos += 4;
-        // [+4:+8] policy_counter_idx u32 LE (#3073)
-        buf[pos..pos + 4].copy_from_slice(&metadata.policy_counter_idx.to_le_bytes());
-        pos += 4;
-        // [+8:+12] inactivity_timeout SECONDS u32 LE (#3227). The metadata
-        // carries ns; the cross-node wire (Go SessionValue.AppTimeout,
-        // SessionSyncRequest.inactivity_timeout) is whole seconds, so convert
-        // ns -> s (saturating). 0 => "use the global per-protocol timeout".
-        let inactivity_secs = match metadata.inactivity_timeout_ns {
-            Some(ns) => u32::try_from(ns / 1_000_000_000).unwrap_or(u32::MAX),
-            None => 0,
-        };
-        buf[pos..pos + 4].copy_from_slice(&inactivity_secs.to_le_bytes());
-        pos += 4;
-
-        // #4565: [+12:+16] the NAT64 translated pool SOURCE (`snat_v4`, 4 raw
-        // IPv4 octets), trailing/length-gated like the #3301 fields. Written for
-        // every open frame (0.0.0.0 when not NAT64); FLAG_NAT64 above is the
-        // semantic gate. This is the ONE piece of the reverse BIB the standby
-        // cannot reconstruct from the synced forward v6 key — the pool source is
-        // chosen by `allocate_source`, not embedded in the key — so it must ride
-        // the wire (the orig v6 src/dst ARE the key; `dst_v4` is the /96 low 32
-        // of the key dst). An old Go decoder length-skips these 4 bytes.
-        let snat_v4 = match (decision.nat.nat64, decision.nat.rewrite_src) {
-            (true, Some(IpAddr::V4(v4))) => v4.octets(),
-            _ => [0u8; 4],
-        };
-        buf[pos..pos + 4].copy_from_slice(&snat_v4);
-        pos += 4;
-
-        // Write header
-        let payload_len = (pos - FRAME_HEADER_SIZE) as u32;
-        write_header(&mut buf, payload_len, MSG_SESSION_OPEN, seq);
-
-        EventFrame {
-            data: buf,
-            len: pos as u16,
-            seq,
-        }
-    }
-
-    /// Encode a SessionClose (type 2) frame -- minimal payload.
-    /// #919/#922: extended with ingress_zone_id + egress_zone_id after the
-    /// flags byte. #3075: widened those two fields from u8 to u16 LE so a stable
-    /// name-hash zone id > 255 round-trips. They are TRAILING fields, so the Go
-    /// decoder length-gates them (a short frame degrades to "no zone ids").
-    pub(crate) fn encode_session_close(
-        seq: u64,
-        key: &SessionKey,
-        owner_rg_id: i32,
-        close_flags: u8,
-        ingress_zone_id: u16,
-        egress_zone_id: u16,
-    ) -> Self {
-        let mut buf = [0u8; 256];
-        let mut pos = FRAME_HEADER_SIZE;
-
-        let is_v6 = key.addr_family == libc::AF_INET6 as u8;
-
-        // [0] AddrFamily
-        buf[pos] = if is_v6 { 6 } else { 4 };
-        pos += 1;
-
-        // [1] Protocol
-        buf[pos] = key.protocol;
-        pos += 1;
-
-        // [2:4] SrcPort
-        buf[pos..pos + 2].copy_from_slice(&key.src_port.to_le_bytes());
-        pos += 2;
-
-        // [4:6] DstPort
-        buf[pos..pos + 2].copy_from_slice(&key.dst_port.to_le_bytes());
-        pos += 2;
-
-        // SrcIP, DstIP
-        pos = write_ip(&mut buf, pos, key.src_ip, is_v6);
-        pos = write_ip(&mut buf, pos, key.dst_ip, is_v6);
-
-        // OwnerRGID i32 LE (#2467: widened from i16, see encode_session_open).
-        buf[pos..pos + 4].copy_from_slice(&owner_rg_id.to_le_bytes());
-        pos += 4;
-
-        // Flags
-        buf[pos] = close_flags;
-        pos += 1;
-
-        // #3075: IngressZoneID u16 LE, EgressZoneID u16 LE (widened from u8).
-        buf[pos..pos + 2].copy_from_slice(&ingress_zone_id.to_le_bytes());
-        pos += 2;
-        buf[pos..pos + 2].copy_from_slice(&egress_zone_id.to_le_bytes());
-        pos += 2;
-
-        let payload_len = (pos - FRAME_HEADER_SIZE) as u32;
-        write_header(&mut buf, payload_len, MSG_SESSION_CLOSE, seq);
-
-        EventFrame {
-            data: buf,
-            len: pos as u16,
-            seq,
-        }
-    }
-
     /// #2460: Encode an RT_FLOW SESSION_CLOSE (type 14) frame.
     ///
     /// Unlike `encode_session_close` (the minimal type-2 HA session-sync
@@ -845,28 +449,6 @@ impl EventFrame {
         }
     }
 
-    /// Encode a DrainComplete (type 8) frame -- header only, no payload.
-    pub(crate) fn encode_drain_complete(seq: u64) -> Self {
-        let mut buf = [0u8; 256];
-        write_header(&mut buf, 0, MSG_DRAIN_COMPLETE, seq);
-        EventFrame {
-            data: buf,
-            len: FRAME_HEADER_SIZE as u16,
-            seq,
-        }
-    }
-
-    /// Encode a FullResync (type 9) frame -- header only, no payload.
-    pub(crate) fn encode_full_resync(seq: u64) -> Self {
-        let mut buf = [0u8; 256];
-        write_header(&mut buf, 0, MSG_FULL_RESYNC, seq);
-        EventFrame {
-            data: buf,
-            len: FRAME_HEADER_SIZE as u16,
-            seq,
-        }
-    }
-
     /// Encode a fixed-size security telemetry event using the existing
     /// RT_FLOW dataplane.Event wire shape consumed by the Go ringbuf parser.
     ///
@@ -934,232 +516,4 @@ impl EventFrame {
             seq,
         }
     }
-
-    /// The raw bytes of this frame (header + payload).
-    pub(crate) fn as_bytes(&self) -> &[u8] {
-        &self.data[..self.len as usize]
-    }
-
-    #[allow(dead_code)]
-    pub(crate) fn dataplane_event_payload(&self) -> Option<&[u8]> {
-        DataplaneEventKind::from_msg_type(self.data[4])?;
-        let payload_len = u32::from_le_bytes(self.data[0..4].try_into().ok()?) as usize;
-        if payload_len != SECURITY_EVENT_PAYLOAD_SIZE {
-            return None;
-        }
-        let end = FRAME_HEADER_SIZE + payload_len;
-        if (self.len as usize) < end {
-            return None;
-        }
-        Some(&self.data[FRAME_HEADER_SIZE..end])
-    }
-
-    pub(super) fn dataplane_event_kind(&self) -> Option<DataplaneEventKind> {
-        DataplaneEventKind::from_msg_type(self.data[4])
-    }
-
-    /// True for the correctness-critical HA session-sync deltas
-    /// (`SessionOpen` / `SessionUpdate` / `SessionClose`). These mutate peer
-    /// session state on failover, so they must not be silently dropped from a
-    /// paused demotion drain window (#2875). Telemetry / RT_FLOW frames
-    /// (deny / screen / filter, RT_FLOW session create+close) return `false`:
-    /// evicting them is a tolerated telemetry loss, not a demotion-correctness
-    /// failure, and must not trigger a spurious FullResync. Same discriminator
-    /// (`data[4]` msg_type) the budget/release path keys on (#2874).
-    pub(super) fn is_session_sync(&self) -> bool {
-        matches!(
-            self.data[4],
-            MSG_SESSION_OPEN | MSG_SESSION_CLOSE | MSG_SESSION_UPDATE
-        )
-    }
-
-    #[allow(dead_code)]
-    pub(crate) fn decode_dataplane_event(&self) -> Option<DataplaneEventPayload> {
-        decode_dataplane_event(self.data[4], self.dataplane_event_payload()?)
-    }
 }
-
-// ---------------------------------------------------------------------------
-// Header / address helpers
-// ---------------------------------------------------------------------------
-
-fn write_header(buf: &mut [u8; 256], payload_len: u32, msg_type: u8, seq: u64) {
-    buf[0..4].copy_from_slice(&payload_len.to_le_bytes());
-    buf[4] = msg_type;
-    // buf[5..8] reserved (already zeroed)
-    buf[8..16].copy_from_slice(&seq.to_le_bytes());
-}
-
-fn write_ip(buf: &mut [u8; 256], pos: usize, ip: IpAddr, is_v6: bool) -> usize {
-    match ip {
-        IpAddr::V4(v4) => {
-            buf[pos..pos + 4].copy_from_slice(&v4.octets());
-            if is_v6 {
-                // pad to 16 bytes if frame is v6 but this particular IP is v4
-                // (shouldn't normally happen, but be safe)
-                pos + 16
-            } else {
-                pos + 4
-            }
-        }
-        IpAddr::V6(v6) => {
-            buf[pos..pos + 16].copy_from_slice(&v6.octets());
-            pos + 16
-        }
-    }
-}
-
-fn write_ip_opt(buf: &mut [u8; 256], pos: usize, ip: Option<IpAddr>, is_v6: bool) -> usize {
-    match ip {
-        Some(addr) => write_ip(buf, pos, addr, is_v6),
-        None => {
-            let size = if is_v6 { 16 } else { 4 };
-            // already zeroed
-            pos + size
-        }
-    }
-}
-
-#[allow(dead_code)]
-fn write_ip_16(buf: &mut [u8; 256], pos: usize, ip: IpAddr) -> usize {
-    match ip {
-        IpAddr::V4(v4) => buf[pos..pos + 4].copy_from_slice(&v4.octets()),
-        IpAddr::V6(v6) => buf[pos..pos + 16].copy_from_slice(&v6.octets()),
-    }
-    pos + 16
-}
-
-#[allow(dead_code)]
-fn write_ip_opt_16(buf: &mut [u8; 256], pos: usize, ip: Option<IpAddr>) -> usize {
-    if let Some(addr) = ip {
-        write_ip_16(buf, pos, addr)
-    } else {
-        pos + 16
-    }
-}
-
-#[allow(dead_code)]
-pub(crate) fn decode_dataplane_event(
-    msg_type: u8,
-    payload: &[u8],
-) -> Option<DataplaneEventPayload> {
-    let frame_kind = DataplaneEventKind::from_msg_type(msg_type)?;
-    if payload.len() != SECURITY_EVENT_PAYLOAD_SIZE {
-        return None;
-    }
-
-    let event_kind = DataplaneEventKind::from_rt_flow_event_type(payload[52])?;
-    if event_kind != frame_kind {
-        return None;
-    }
-    let wire_af = payload[55];
-    if wire_af != RT_FLOW_AF_INET && wire_af != RT_FLOW_AF_INET6 {
-        return None;
-    }
-    let policy_or_reason_id = u32::from_le_bytes(payload[44..48].try_into().ok()?);
-
-    Some(DataplaneEventPayload {
-        kind: event_kind,
-        addr_family: if wire_af == RT_FLOW_AF_INET6 {
-            libc::AF_INET6 as u8
-        } else {
-            libc::AF_INET as u8
-        },
-        protocol: payload[53],
-        action: payload[54],
-        src_port: u16::from_be_bytes(payload[40..42].try_into().ok()?),
-        dst_port: u16::from_be_bytes(payload[42..44].try_into().ok()?),
-        nat_src_port: u16::from_be_bytes(payload[104..106].try_into().ok()?),
-        nat_dst_port: u16::from_be_bytes(payload[106..108].try_into().ok()?),
-        ingress_zone_id: u16::from_le_bytes(payload[48..50].try_into().ok()?),
-        egress_zone_id: u16::from_le_bytes(payload[50..52].try_into().ok()?),
-        ingress_ifindex: i32::from_le_bytes(payload[128..132].try_into().ok()?),
-        rule_id: u32::from_le_bytes(payload[56..60].try_into().ok()?),
-        term_id: u32::from_le_bytes(payload[60..64].try_into().ok()?),
-        owner_rg_id: i16::from_le_bytes(payload[64..66].try_into().ok()?),
-        policy_id: if event_kind == DataplaneEventKind::PolicyDeny {
-            policy_or_reason_id
-        } else {
-            0
-        },
-        reason: payload[134],
-        application_id: u16::from_le_bytes(payload[132..134].try_into().ok()?),
-        filter_id: if event_kind == DataplaneEventKind::FilterLog {
-            policy_or_reason_id
-        } else {
-            0
-        },
-        screen_id: if event_kind == DataplaneEventKind::ScreenDrop {
-            policy_or_reason_id
-        } else {
-            0
-        },
-        timestamp_ns: u64::from_le_bytes(payload[0..8].try_into().ok()?),
-        src_ip: read_ip_16(&payload[8..24], wire_af)?,
-        dst_ip: read_ip_16(&payload[24..40], wire_af)?,
-        nat_src_ip: read_nonzero_ip_16(&payload[72..88], wire_af),
-        nat_dst_ip: read_nonzero_ip_16(&payload[88..104], wire_af),
-    })
-}
-
-#[allow(dead_code)]
-fn read_ip_16(bytes: &[u8], wire_af: u8) -> Option<IpAddr> {
-    match wire_af {
-        RT_FLOW_AF_INET => Some(IpAddr::from(<[u8; 4]>::try_from(&bytes[..4]).ok()?)),
-        RT_FLOW_AF_INET6 => Some(IpAddr::from(<[u8; 16]>::try_from(&bytes[..16]).ok()?)),
-        _ => None,
-    }
-}
-
-#[allow(dead_code)]
-fn read_nonzero_ip_16(bytes: &[u8], wire_af: u8) -> Option<IpAddr> {
-    if bytes.iter().all(|b| *b == 0) {
-        return None;
-    }
-    read_ip_16(bytes, wire_af)
-}
-
-#[allow(dead_code)]
-fn rt_flow_addr_family(addr_family: u8, src_ip: IpAddr) -> u8 {
-    if addr_family == libc::AF_INET6 as u8 || matches!(src_ip, IpAddr::V6(_)) {
-        RT_FLOW_AF_INET6
-    } else {
-        RT_FLOW_AF_INET
-    }
-}
-
-fn encode_disposition(d: ForwardingDisposition) -> u8 {
-    match d {
-        ForwardingDisposition::ForwardCandidate => DISP_FORWARD_CANDIDATE,
-        ForwardingDisposition::LocalDelivery => DISP_LOCAL_DELIVERY,
-        ForwardingDisposition::FabricRedirect => DISP_FABRIC_REDIRECT,
-        ForwardingDisposition::PolicyDenied => DISP_POLICY_DENIED,
-        ForwardingDisposition::NoRoute => DISP_NO_ROUTE,
-        ForwardingDisposition::MissingNeighbor => DISP_MISSING_NEIGHBOR,
-        ForwardingDisposition::HAInactive => DISP_HA_INACTIVE,
-        ForwardingDisposition::DiscardRoute => DISP_DISCARD_ROUTE,
-        ForwardingDisposition::NextTableUnsupported => DISP_NEXT_TABLE_UNSUPPORTED,
-    }
-}
-
-/// Compute the close flags byte from a SessionDelta.
-pub(crate) fn close_flags(delta: &SessionDelta) -> u8 {
-    let mut flags: u8 = 0;
-    if delta.fabric_redirect_sync
-        || delta.decision.resolution.disposition == ForwardingDisposition::FabricRedirect
-    {
-        flags |= FLAG_FABRIC_REDIRECT;
-    }
-    if delta.metadata.fabric_ingress {
-        flags |= FLAG_FABRIC_INGRESS;
-    }
-    flags
-}
-
-// ---------------------------------------------------------------------------
-// Tests
-// ---------------------------------------------------------------------------
-
-#[cfg(test)]
-#[path = "codec_tests.rs"]
-mod tests;
